requirements.txt
fastapi==0.85.1
uvicorn==0.19.0
requests==2.28.1

# for database
SQLAlchemy==1.4.42
psycopg2==2.9.5

# for loading environment variables
python-dotenv==0.21.0

# for tests
pytest==7.2.0

# for visualization
pandas==1.5.1
numpy==1.23.4
matplotlib==3.6.0
seaborn==0.12.1
